Ok...it's working (HW1, HW2, AllFunctions)
It appears to be working.

Switched sign for bot/top forcing terms, making sure they are added to the forcing term (assuming forcing term is on the same side of the equation as the rest of the parameters). In HW1 for example, where the forcing term is on the opposite side, I had to multiply the forcing term by -1 to get the proper result.

In HW2, added mixed layer shifting horizontal line on the animation and also a line delineating the mixed layer in the contour plots.